+++ examples/header.rs
@@ -8,9 +8,9 @@ struct Plugin;
 
 impl Guest for Plugin {
     fn handle_request(&self, request: Request, _response: Response) -> (bool, i32) {
-        request
-            .header()
-            .add(&Bytes::from("X-Foo"), &Bytes::from("Bar"));
+        let header = request.header();
+        header.add(&Bytes::from("X-Foo"), &Bytes::from("foo"));
+        header.add(b"X-Bar", b"bar");
         (true, 0)
     }
 }